+++ src/Events/Event.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Accu\Postmen\Events;
+
+interface Event
+{
+    public function data();
+}